Only consider purely non-alpha words for comments
Closes #6.

Previously, any word (by which we mean a whitespace-separated token)
that had *any* non-alpha characters in it would be considered a comment
prefix.

We've now swapped this condition, such that only words which consist of
only non-alpha characters are considered comment prefixes.

This eliminates words like "I'm", "So" or "Hi!" from being considered
comment prefixes.